+++ changelog.md
@@ -9,6 +9,8 @@ Codion Change Log
 - BlobColumnDefinition removed, ColumnDefinition.lazy() added. Default behaviour for byte array columns being lazy and hidden removed. Related changes.
 ### is.codion.framework.db
 - EntityConnection.readBlob() and writeBlob() removed.
+### is.codion.framework.model
+- EntityTableModel.attributes() added, specifying the attributes to include when selecting. Implemented in SwingEntityTableModel, replacing overridable attributes() method. Related refactoring.
 
 ## 0.17.18
 - Text.padString() deprecated for removal along with Text.ALIGNMENT, Text.leftPad() and rightPad() added.